Changed the interpolation interface to stripack to better reflect the return codes
The interpolation routine in stripack returns 0 for a successful
interpolation, 1 for a successful extrapolation and negative values for
errors. The previous error checking process was killing any case with
extrapolation as well as error conditions. I instead return an
information vector that allows checking of individual points and prints
a warning if one or more errors has been flagged on the way.